backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
psycopg[binary]>=3.1
celery>=5.4
redis>=5.0
pydantic>=2.7
requests>=2.32
beautifulsoup4>=4.12
pytest>=8.0
pytest-django>=4.8
